Separated custom audio player class into new file
Called HybridAudioPlayer, can't think of a better name...
Also added some documentation regarding audio handling in the microbe
class.
