+++ src/microbe_stage/Microbe.cs
@@ -24,7 +24,7 @@ public partial class Microbe : RigidBody, ISpawned, IProcessable, IMicrobeAI, IS
     /// </summary>
     public Vector3 MovementDirection = new Vector3(0, 0, 0);
 
-    private AudioPlayer engulfAudio;
+    private HybridAudioPlayer engulfAudio;
     private AudioStreamPlayer3D bindingAudio;
     private AudioStreamPlayer3D movementAudio;
     private List<AudioStreamPlayer3D> otherAudioPlayers = new List<AudioStreamPlayer3D>();
@@ -184,7 +184,7 @@ public partial class Microbe : RigidBody, ISpawned, IProcessable, IMicrobeAI, IS
 
         cellBurstEffectScene = GD.Load<PackedScene>("res://src/microbe_stage/particles/CellBurstEffect.tscn");
 
-        engulfAudio = new AudioPlayer
+        engulfAudio = new HybridAudioPlayer
         {
             Stream = GD.Load<AudioStream>("res://assets/sounds/soundeffects/engulfment.ogg"),
             Positional = !IsPlayerMicrobe,
@@ -192,6 +192,12 @@ public partial class Microbe : RigidBody, ISpawned, IProcessable, IMicrobeAI, IS
         };
 
         AddChild(engulfAudio);
+
+        // You may notice that there are two separate ways that an audio is played in this class:
+        // using pre-existing audio node e.g "bindingAudio", "movementAudio" and through method e.g "PlaySoundEffect",
+        // "PlayNonPositionalSoundEffect". The former is approach best used to play looping sounds with more control
+        // to the audio player while the latter is convenient for dynamic and various short one-time sound effects
+        // in expense of lesser audio player control.
 
         if (IsPlayerMicrobe)
         {
@@ -719,114 +725,5 @@ public partial class Microbe : RigidBody, ISpawned, IProcessable, IMicrobeAI, IS
         }
 
         cachedHexCountDirty = false;
-    }
-
-    /// <summary>
-    ///   Provides extra level of abstraction to allow simultaneous switching between 3D positional and non positional
-    ///   audio players in a single Spatial-derived class.
-    /// </summary>
-    /// <remarks>
-    ///   <para>
-    ///     Useful in cases where the playing of an audio stream must be able to fulfill both of these conditions.
-    ///   </para>
-    /// </remarks>
-    private class AudioPlayer : Spatial
-    {
-        private AudioStreamPlayer3D player3D;
-        private AudioStreamPlayer playerNonPositional;
-
-        private float volume;
-
-        public AudioPlayer()
-        {
-            player3D = new AudioStreamPlayer3D();
-            playerNonPositional = new AudioStreamPlayer();
-
-            AddChild(player3D);
-            AddChild(playerNonPositional);
-
-            Volume = 1.0f;
-        }
-
-        public bool Positional { get; set; }
-
-        public AudioStream Stream
-        {
-            get => Positional ? player3D.Stream : playerNonPositional.Stream;
-            set
-            {
-                if (Positional)
-                {
-                    player3D.Stream = value;
-                }
-                else
-                {
-                    playerNonPositional.Stream = value;
-                }
-            }
-        }
-
-        public bool Playing => Positional ? player3D.Playing : playerNonPositional.Playing;
-
-        /// <summary>
-        ///   Volume in linear scale.
-        /// </summary>
-        public float Volume
-        {
-            get => volume;
-            set
-            {
-                volume = Mathf.Clamp(value, 0, 1);
-
-                if (Positional)
-                {
-                    player3D.UnitDb = GD.Linear2Db(volume);
-                }
-                else
-                {
-                    playerNonPositional.VolumeDb = GD.Linear2Db(volume);
-                }
-            }
-        }
-
-        public string Bus
-        {
-            get => Positional ? player3D.Bus : playerNonPositional.Bus;
-            set
-            {
-                if (Positional)
-                {
-                    player3D.Bus = value;
-                }
-                else
-                {
-                    playerNonPositional.Bus = value;
-                }
-            }
-        }
-
-        public void Play(float fromPosition = 0)
-        {
-            if (Positional)
-            {
-                player3D.Play(fromPosition);
-            }
-            else
-            {
-                playerNonPositional.Play(fromPosition);
-            }
-        }
-
-        public void Stop()
-        {
-            if (Positional)
-            {
-                player3D.Stop();
-            }
-            else
-            {
-                playerNonPositional.Stop();
-            }
-        }
     }
 }
